mydjango/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-filter==24.2
django-simple-history==3.5.0
djangorestframework==3.15.1
sqlparse==0.5.0
typing_extensions==4.11.0
